fix: line 47 in main: add player x and y start pos
