+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: bgales <bgales@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/18 18:19:22 by bgales            #+#    #+#             */
-/*   Updated: 2023/04/26 15:40:52 by bgales           ###   ########.fr       */
+/*   Updated: 2023/06/08 22:20:21 by bgales           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,4 +31,11 @@
 # include "game.h"
 
 char	*get_next_line(int fd);
+int		create_trgb(int t, int r, int g, int b);
+void 	draw(t_data img, int x, int draw_start, int draw_end, int color);
+int		key_press_hook(int keycode, void *params);
+double	time_calculator(void);
+void	game_loop(t_game game, t_data img, double pos_x, double pos_y, double dir_x, double dir_y, double plane_x, double plane_y, char **map_ig);
+int		exit_game(void);
+
 #endif
