+++ lib/libft/ft_isdigit.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_isdigit.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ctardy <ctardy@student.42nice.fr>          +#+  +:+       +#+        */
+/*   By: bgales <bgales@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/20 12:52:31 by ctardy            #+#    #+#             */
-/*   Updated: 2021/10/28 15:31:50 by ctardy           ###   ########.fr       */
+/*   Updated: 2023/06/08 22:13:53 by bgales           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,6 @@
 
 int	ft_isdigit(int c)
 {
-	int	i;
-
-	i = 0;
 	if (c >= 48 && c <= 57)
 		return (1);
 	return (0);
