+++ lib/libft/ft_strlen.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ctardy <ctardy@student.42nice.fr>          +#+  +:+       +#+        */
+/*   By: bgales <bgales@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/20 13:05:13 by ctardy            #+#    #+#             */
-/*   Updated: 2021/10/25 19:59:39 by ctardy           ###   ########.fr       */
+/*   Updated: 2023/06/09 09:17:35 by bgales           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const	char *str)
+int	ft_strlen(const	char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i])
