+++ lib/libft/ft_substr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ctardy <ctardy@student.42nice.fr>          +#+  +:+       +#+        */
+/*   By: bgales <bgales@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/24 18:24:41 by ctardy            #+#    #+#             */
-/*   Updated: 2021/11/01 18:30:58 by ctardy           ###   ########.fr       */
+/*   Updated: 2023/06/09 09:19:34 by bgales           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,9 +20,9 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	i = 0;
 	if (s == NULL)
 		return (NULL);
-	if (len > ft_strlen(s + start))
+	if ((int)len > ft_strlen(s + start))
 		len = ft_strlen(s) - start;
-	if (start > ft_strlen(s))
+	if ((int)start > ft_strlen(s))
 		return (ft_strdup(""));
 	inter = malloc(sizeof(char) * (len + 1));
 	if (inter == NULL)
