+++ src/game/raytracing.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   raytracing.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ctardy <ctardy@student.42nice.fr>          +#+  +:+       +#+        */
+/*   By: bgales <bgales@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/18 17:16:58 by ctardy            #+#    #+#             */
-/*   Updated: 2023/06/08 20:11:41 by ctardy           ###   ########.fr       */
+/*   Updated: 2023/06/09 09:47:38 by bgales           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -160,10 +160,10 @@ void game_loop(t_game game, t_data img, double pos_x, double pos_y, double dir_x
           			map_y += step_y;
           			side = 1;
         		}
-
         //Check if ray has hit a wall
-        		if (map_ig[map_x][map_y] > '0')
+        		if (map_ig[map_y][map_x] > '0')
 					hit = 1;
+
       		}
 
       //Calculate distance projected on camera direction (Euclidean distance would give fisheye effect!)
@@ -183,8 +183,10 @@ void game_loop(t_game game, t_data img, double pos_x, double pos_y, double dir_x
     		if (draw_end >= h)
 				draw_end = h - 1;
 
+				printf("x = %d, y = %d, hit = %d\n", map_x, map_y, hit);
      	 //choose wall color
 			int color;
+			printf("%d %d\n", map_x, map_y);
 			color = color_select(map_ig[map_x][map_y]);
 
       //give x and y sides different brightness
@@ -306,63 +308,6 @@ int key_press_hook(int keycode, void *params)
 	game_loop(*game, game->imgig, game->numig.pos_x, game->numig.pos_y, game->numig.dir_x, game->numig.dir_y, game->numig.plane_x, game->numig.plane_y, game->map_ig);
 	mlx_put_image_to_window(game->mlx, game->window, game->imgig.img, 0, 0); // redraw();
 	return 0;
-}
-
-int main (int argc, char **argv)
-{
-
-	/***********************************BEN PART*************************************************/
-
-	t_parse	*parse;
-	char	**cub_file;
-
-	parse = malloc(sizeof(t_parse));
-	if (argc != 2)
-	{
-		printf("Wrong argument format");
-		exit (0);
-	}
-	parser(&parse, cub_file, argv[1]);
-
-	char **map = parse->map;
-	while (*map)
-		puts(*(map++));
-
-	
-	exit(0);
-
-
-
-	/***********************************CHARLY PART*************************************************/
-
-
-	(void)argc;
-	(void)argv;
-	t_game	game;
-
-	ft_memset(&game, 0, sizeof(t_game));
-	game.map_ig = parse->map;
-
-	game.numig.pos_x = 14, game.numig.pos_y = 14;  //x and y start position
-	game.numig.dir_x = -1, game.numig.dir_y = 0; //initial direction vector
-	game.numig.plane_x = 0, game.numig.plane_y = 0.66; //the 2d raycaster version of camera plane
-	game.numig.time = 0; //time of current frame
-	game.numig.start = time_calculator(); //time of current frame
-	game.numig.old_time = 0; //time of previous frame
-	game.numig.old_time = game.numig.time;
-    game.numig.time = time_calculator() - game.numig.start;
-
-	game.mlx = mlx_init();
-	game.window = mlx_new_window(game.mlx, 1024, 720, "cub3D");
-	game.imgig.img = mlx_new_image(game.mlx, 1024, 720);
-	game.imgig.addr = mlx_get_data_addr(game.imgig.img, &game.imgig.bits_per_pixel, &game.imgig.line_length, &game.imgig.endian);
-
-	game_loop(game, game.imgig, game.numig.pos_x, game.numig.pos_y, game.numig.dir_x, game.numig.dir_y, game.numig.plane_x, game.numig.plane_y, game.map_ig);
-
-	mlx_put_image_to_window(game.mlx, game.window, game.imgig.img, 0, 0); // redraw();
-	mlx_hook(game.window, 17, (1L << 17), exit_game, &game);
- 	mlx_hook(game.window, 2, (1L << 0), key_press_hook, &game); //readKeys();
-	mlx_loop(game.mlx);
 }
