+++ src/get/get_info.c
@@ -6,7 +6,7 @@
 /*   By: bgales <bgales@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/21 12:24:42 by bgales            #+#    #+#             */
-/*   Updated: 2023/06/08 16:01:58 by bgales           ###   ########.fr       */
+/*   Updated: 2023/06/08 23:14:34 by bgales           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
