+++ src/parsing/parser.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parser.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ctardy <ctardy@student.42nice.fr>          +#+  +:+       +#+        */
+/*   By: bgales <bgales@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/20 12:20:59 by bgales            #+#    #+#             */
-/*   Updated: 2023/06/08 20:08:30 by ctardy           ###   ########.fr       */
+/*   Updated: 2023/06/08 23:18:53 by bgales           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
