+++ src/utils/parse_utils.c
@@ -6,7 +6,7 @@
 /*   By: bgales <bgales@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/13 16:30:27 by bgales            #+#    #+#             */
-/*   Updated: 2023/06/08 15:54:42 by bgales           ###   ########.fr       */
+/*   Updated: 2023/06/08 23:06:02 by bgales           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
